feat: introduce CategoryPieChart and WalletPieChart components for enhanced transaction visualization
