+++ src/pages/Home/components/CategorySection/index.tsx
@@ -1,20 +1,9 @@
 import { useNavigation } from '@react-navigation/native';
 import React, { useContext } from 'react';
-import TransactionPieChart from '../../../../components/TransactionPieChart';
+import { CategoryPieChart } from '../../../../components/TransactionPieChart';
 import Text from '../../../../components/Text';
 import AppContext from '../../../../contexts/AppContext';
-import { Transaction } from '../../../../models';
-import { getCategoryById, getDefaultCategoryByType } from '../../../../utils/category';
 import { SectionContainer, SectionHeader, SeeMoreButton } from '../commonStyles';
-
-const resolveExpenseCategory = (t: Transaction) =>
-  getCategoryById(t.categoryId) ?? getDefaultCategoryByType('EXPENSE');
-
-const getExpenseSegmentId = (t: Transaction) => resolveExpenseCategory(t).id;
-
-const getExpenseSegmentName = (t: Transaction) => resolveExpenseCategory(t).name;
-
-const getExpenseSegmentColor = (t: Transaction) => resolveExpenseCategory(t).color;
 
 const CategorySection: React.FC = () => {
   const { expenseTransactions } = useContext(AppContext);
@@ -27,12 +16,10 @@ const CategorySection: React.FC = () => {
         <Text typography="title">Despesas por categoria</Text>
         <SeeMoreButton text="Ver mais" onPress={() => navigation.navigate('insights')} />
       </SectionHeader>
-      <TransactionPieChart
+      <CategoryPieChart
+        type="EXPENSE"
         transactions={expenseTransactions}
         variant="inline"
-        getSegmentId={getExpenseSegmentId}
-        getSegmentName={getExpenseSegmentName}
-        getSegmentColor={getExpenseSegmentColor}
       />
     </SectionContainer>
   );
